# persist: hook up an initial implementation of compaction