--- a/src/persist-client/src/impl/machine.rs
+++ b/src/persist-client/src/impl/machine.rs
@@ -28,7 +28,7 @@ use crate::error::InvalidUsage;
 use crate::r#impl::metrics::{
     CmdMetrics, Metrics, MetricsRetryStream, RetriesMetrics, RetryMetrics,
 };
-use crate::r#impl::state::{ReadCapability, Since, State, StateCollections, Upper};
+use crate::r#impl::state::{HollowBatch, ReadCapability, Since, State, StateCollections, Upper};
 use crate::read::ReaderId;
 use crate::ShardId;
 
@@ -116,14 +116,13 @@ where
 
     pub async fn compare_and_append(
         &mut self,
-        keys: &[String],
-        desc: &Description<T>,
+        batch: &HollowBatch<T>,
     ) -> Result<Result<Result<SeqNo, Upper<T>>, InvalidUsage<T>>, Indeterminate> {
         let metrics = Arc::clone(&self.metrics);
         loop {
             let (seqno, res) = self
                 .apply_unbatched_cmd(&metrics.cmds.compare_and_append, |_, state| {
-                    state.compare_and_append(keys, desc)
+                    state.compare_and_append(batch)
                 })
                 .await?;
 
@@ -143,7 +142,7 @@ where
 
                     // We tried to to a compare_and_append with the wrong
                     // expected upper, that won't work.
-                    if &current_upper != desc.lower() {
+                    if &current_upper != batch.desc.lower() {
                         return Ok(Ok(Err(Upper(current_upper))));
                     } else {
                         // The upper stored in state was outdated. Retry after
@@ -243,14 +242,11 @@ where
         }
     }
 
-    pub async fn next_listen_batch(
-        &mut self,
-        frontier: &Antichain<T>,
-    ) -> (Vec<String>, Description<T>) {
+    pub async fn next_listen_batch(&mut self, frontier: &Antichain<T>) -> HollowBatch<T> {
         let mut retry: Option<MetricsRetryStream> = None;
         loop {
-            if let Some((keys, desc)) = self.state.next_listen_batch(frontier) {
-                return (keys.to_owned(), desc.clone());
+            if let Some(b) = self.state.next_listen_batch(frontier) {
+                return b;
             }
             // Only sleep after the first fetch, because the first time through
             // maybe our state was just out of date.