--- a/src/persist-client/src/lib.rs
+++ b/src/persist-client/src/lib.rs
@@ -52,6 +52,7 @@ pub(crate) mod r#impl {
     pub mod machine;
     pub mod metrics;
     pub mod state;
+    pub mod trace;
 }
 
 // TODO: Remove this in favor of making it possible for all PersistClients to be