--- a/src/persist-client/tests/trace/since_upper
+++ b/src/persist-client/tests/trace/since_upper
@@ -7,7 +7,7 @@
 # the Business Source License, use of this software will be governed
 # by the Apache License, Version 2.0.
 
-# since and upper start a minimum ts
+# since and upper start at minimum ts
 since-upper
 ----
 [0][0]