… Cargo.lock
… src/persist-client/Cargo.toml
… src/persist-client/examples/open_loop.rs
… src/persist-client/src/impl/metrics.rs
… src/persist-client/src/impl/trace.rs
… src/persist-client/tests/machine/compaction
… src/persist-client/tests/machine/compaction_empty
… src/persist-client/tests/machine/compaction_multipart
… src/persist-client/tests/machine/compaction_truncate
… src/persist-client/tests/machine/regression_listen_compaction
… src/persist-client/tests/machine/truncate
… src/persist-client/tests/trace/compaction
… src/persist-client/tests/trace/empty_batch_optimization
… src/persist/src/cfg.rs
… src/persist/src/indexed/background.rs
… src/persist/src/indexed/columnar/parquet.rs
… src/persist/src/indexed/encoding.rs