See the comments in trace.rs for an overview of how this is set up.

Compaction tests can quickly become a maintenance burden as it's
difficult to balance invariant verification against needing lots of
diffs when small compaction heuristics change. Invest early in
datadriven tests (the same style and underlying library as testdrive).

In production initially disabled by default, but turned on by the
`MZ_PERSIST_COMPACTION_ENABLED` env var for easy testing. In test,
enabled by default.
### Motivation

  * This PR adds a known-desirable feature.

### Tips for reviewer

I'd recommend taking a quick scan of the entire PR's diff to get a high level picture (maybe start with the comment in trace.rs) and then going back commit by commit.

### Testing

- [x] This PR has adequate test coverage / QA involvement has been duly considered.

### Release notes

This PR includes the following [user-facing behavior changes](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/guide-changes.md#what-changes-require-a-release-note):

  - N/A