@danhhz 2022-06-17
fyi @aljoscha @pH14

I had been hoping to get this out for review this week, but I'm coming up just shy on time (and focus). no need to review yet but absolutely do feel free to poke around and start leaving comments if you like

one big missing piece here that I owe is a justification for why I've forked Spine instead of using it directly. there are pros/cons to each and it'll take a bit to type up

I also have a couple final WIPs left and some testing to make sure that this doesn't mess up dev/CI even while disabled. I'd also like to make sure it doesn't noticeably regress the benchmarks, in particular open-loop