@danhhz 2022-06-21
> some testing to make sure that this doesn't mess up dev/CI even while disabled

okay well that answers that. looks like some CI jobs (most notably testdrive) are _significantly_ slower with this PR in its current state, so I'll have to figure out what's going on there and a way to address it before this merges