@danhhz 2022-06-21
my initial guess for the slowdown was that we're doing non-trivial work in decode now (ingesting batches 1:1 and letting the Spine decide where they end up), so put together a quick benchmark for state encode and decode. it's slower but I can't imagine this is enough to explain testdrive given the magnitude of the numbers. this would be a nice benchmark to merge, but it would require making quite a few things pub that currently are not, so I think it deserves it's own PR if we decide it's worth the work

i don't immediately have a second guess of what is causing the slowdown. i suppose my next steps are to read over the code again with a critical eye for that and/or running open loop (the latter I wanted to do anyway)

before (on my laptop)
```
MZ_PERSIST_RECORD_COUNT_SMALL=64 MZ_PERSIST_RECORD_SIZE_BYTES_SMALL=1024 MZ_PERSIST_BATCH_MAX_COUNT_SMALL=8
plumbing/encode_state/state/64KiB
                        time:   [1.6603 us 1.6613 us 1.6626 us]
Found 5 outliers among 100 measurements (5.00%)
  3 (3.00%) high mild
  2 (2.00%) high severe

plumbing/decode_state/state/64KiB
                        time:   [3.6962 us 3.7005 us 3.7057 us]
Found 10 outliers among 100 measurements (10.00%)
  10 (10.00%) high severe
```

after (also on my laptop)
```
MZ_PERSIST_RECORD_COUNT_SMALL=64 MZ_PERSIST_RECORD_SIZE_BYTES_SMALL=1024 MZ_PERSIST_BATCH_MAX_COUNT_SMALL=8
plumbing/encode_state/state/64KiB
                        time:   [1.7780 us 1.7793 us 1.7807 us]
                        change: [+7.2663% +7.4660% +7.6840%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 9 outliers among 100 measurements (9.00%)
  3 (3.00%) high mild
  6 (6.00%) high severe

plumbing/decode_state/state/64KiB
                        time:   [8.1047 us 8.1108 us 8.1183 us]
                        change: [+119.15% +119.45% +119.76%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 6 outliers among 100 measurements (6.00%)
  1 (1.00%) low mild
  5 (5.00%) high severe
```