@danhhz 2022-06-21
yup still repros. I'm a little stumped because I have a bunch of theories for things that might have gotten slower (added "empty batch optimization is not as comprehensive as before"), but so far nothing to explain why "as of ... not yet available". failed theories for asof are:
- bug in new upper calculation (should readily repro locally)
- slowness is causing starvation so the upper never gets updated (feels like more should be broken if it's bad enough to consistently break github-5983.td, which is as simple as it gets contention-wise)

about the only idea I have left atm is to hack together a change that speeds up decode by directly constructing spine's inner state. that should get it back to the same range it was before, but it's a long shot. I might instead take the rest of my day to circle back to the commentary I meant to write for this and let the testdrive problem mull in the background