@danhhz 2022-06-21
As promised, here are the pros/cons that I have thought of for forking Spine vs using it directly:

Forking Spine
- (1) pro: implements `Send` and so works with persist's existing code structure. the dd Spine is not `Send` because of some internal `Rc`s and so would require us to switch to a model where we communicate with `Machine` over channels. in the long term, that may be a beneficial thing to do anyway (e.g. could help reduce CaS expectation failures when a reader+writer are in the same process), but I don't want to block compaction on it. so in the short term, this is ultimately why I forked in this PR
- (2) pro: allows us to mess directly with the internals of Spine. e.g. we could serialize/deserialize the spine internal levels directly (vs decode working by reintroducing all the batches and letting them advance though the spine logic as normal). if we have to speed up decode (as I've wondered in comments above) this is basically our only knob, so I'd like to keep it in the back pocket
- (3) pro: being able to litter in printlns and such during development was pretty invaluable to getting this all working. this is also possible if we used dd spine by pointing a Cargo.toml override at a local dd checkout, but it's certainly easier
- (4) con: a vague sense of dissatisfaction at the duplication
- (5) con: the work of keeping the fork's changes minimal
- (6) con: having to explain why we forked it when engineers are curious

Using DD Spine Directly
- (7) pro: if we could figure out a way to hook up `Cursor` (I have ideas), an actual Spine impl plus some tweaks to Snapshot+Listen gets us close to an `Arrangement` that's based by persist. the latency of accesses would be wildly different vs the normal in-mem Arrangement, so it's unclear if this is actually useful. it's certainly a very cool idea though
- (8) pro: a vague sense of satisfaction
- (9) pro: picking up Spine improvements for free (for reference: there have been 2 PRs in the last year, one to improve the generics that we've tossed in the fork and one with a minor alloc reuse)
- (10) con: it'd be tough to discover new merge requests. we'd have to do it via `map_batches` I think?

in summary: I think we'll need to periodically re-evaluate whether forking is the right choice, but I feel it is while we're in the rapid development phase. also, I'm reasonably certain that (modulo `Send`-ness) the `Trace` interface here (the only thing exposed to the rest of the crate) is possible to implement without the fork, which would make it easier to reverse this decision if the pro/con math changes (1 and 7 are the ones I'm most looking at)